02 with authentication

+++ controllers/feed.controller.js
@@ -1,17 +1,18 @@
 const { validationResult } = require("express-validator");
-const Post = require("../models/post.model");
 const fs = require("fs");
 const path = require("path");
+const Post = require("../models/post.model");
+const User = require("../models/user.model");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
   const pageSize = 2;
   let totalPosts;
-  Post.find()
+  Post.find({ creator: req.userId })
     .countDocuments()
     .then((count) => {
       totalPosts = count;
-      return Post.find()
+      return Post.find({ creator: req.userId })
         .skip((currentPage - 1) * pageSize)
         .limit(pageSize);
     })
@@ -66,15 +67,22 @@ exports.addPost = (req, res, next) => {
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: "Hardik P",
-    },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((user) => {
       res.status(200).json({
-        post: result,
+        message: "Post created",
+        post: post,
+        creator: { _id: user._id, name: user.name },
       });
     })
     .catch((err) => {
@@ -86,7 +94,6 @@ exports.addPost = (req, res, next) => {
 };
 
 exports.updatePost = (req, res, next) => {
-  console.log("here");
   const errors = validationResult(req);
   if (errors.errors && errors.errors.length > 0) {
     const error = new Error("Invalid data");
@@ -115,6 +122,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -145,11 +159,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findOneAndDelete(postId);
     })
     .then(() => {
-      res.status(200).json({ message: "Product Deleted successfully..." });
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      // This will clear the post id in users table
+      return user.posts.pull(postId);
+    })
+    .then(() => {
+      res.status(200).json({ message: "Post Deleted successfully..." });
     })
     .catch((err) => {
       if (!err.statusCode) {
